Add scroller property to vaadin-adaptive-overlay.
Adding the scroller property will make <iron-list> to use the
returned element as the scrolling target. By default, <iron-list>
uses itself as a target.

<iron-list> calculates the viewport height
from the height of the scroller element, and in IE11, the height is
calculated incorrectly when <iron-list> itself is used as a target.

Wrong viewportSize affects both calculations done in <vaadin-combo-box>
for scrolling and scrollToIndex() in <iron-list>.

For touch scrolling, the scroller element has to be <iron-list> in order
for scrolling to work properly.