Fixes #48 items text overlap above the underline by correcting the vertical coordinate of the dropdown
This fix introduces a small correction for the overlay's dropdown vertical offset. The offset is shifted by a small value, to make the vertical coordinate of the dropdown equal interger number of device pixels. This resolves rounding problems with Safari rendering engine, making the combo-box items text to always render inside the dropdown, behind the underline.