Turned off the iron-dropdown alignment
After enabling the body scroll, the dropdown must have `position: absolute`
placement instead of `position: fixed`. But simply switching the styles was not
enough, because the `<iron-dropdown/>` alignment code becomes broken and
useless. No more need to calculate the viewport coordinates of the dropdown.

This fix binds `positionTarget` property of the `iron-dropdown` to `null`, which
effectively turns off its coordinate calculations. Instead the
`_updateOverlayPosition` method was added in order to manually set the top
offset of the dropdown.