Replace iron-dropdown with vaadin-dropdown.
<iron-dropdown> and iron-overlay-behavior uses position:fixed for the overlay,
which causes issues with iOS Safari and enabling body scrolling. Setting
position to absolute helps, but might cause issues when overlay is nested
inside a parent element that for example has overflow:hidden.

Placing the overlay directly under document body is what most overlay widgets
seem to do, including the Vaadin Framework components.

In order to achieve this, the most feasible option is to drop
iron-overlay-behavior (and therefore iron-dropdown) and use only
iron-fit-behavior and implement the few needed features ourselves.