Focus overlay explicitly when clicked on.
We are effectively blurring the input field when overlay is tapped in order to
hide the virtual keyboard on touch devices.

However, at least on WIN10 IE11 and Edge the focus is moved to document.body
when the input is blurred, forcing the overlay to be closed and selected value
to be reverted.

Setting the focus to overlay (#scroller) now that it has tabindex fixes this,
and also makes the code reflect better on what we're trying to achieve.