Allow the developer to replace the button icons
Fixes #196

Remove the dependency to iron-icon(s), to make it easy to replace the
icons with CSS mixins by overriding the background-image, instead of
changing the attribute of the iron-icon element.

Revert the toggle button icon back to the dropdown arrow at the same
time, as the “menu” icon was a failed experiment.