Change toggle icon back to dropdown arrow
We failed to gather enough feedback about the change to the “menu”
icon, so it’s safer to revert back at this point.

Align the icons vertically exactly like in paper-dropdown-menu at the
same time (shift upwards slightly), and align the icon color as well.

#196 provides a way to change the icon, if in the future we feel like
we would want to use another icon in our own theme for example.