Fix the notifyPath call to have the explicit `undefined` value argument
The second and the third arguments in Polymer.Base.notifyPath are now optional since
Polymer v1.5.0. Because of that, the behavior when called with a missing argument has
changed. We have to use explicit `undefined` value argument from now on.

More about the change: https://github.com/Polymer/polymer/pull/3473