Postpone overlay stamping to first opening (#750)
This is a performance improvement.

When attaching a lot of combo boxes at once, the initial render can take a lot of time. Most of the time is spent on executing logic related to the overlay. This functionality is not needed until the overlay is opened, and it is wasted effort for any combo box on the page that the user doesn't open.

As a trade-off the first opening takes a bit longer. But the bigger problem was the huge delay caused by initializing a lot of combo boxes at once.

Fixes #748