Don't request empty loaded page again
Previously, combobox didn't request a page again if it was already
loaded, but this logic did not apply if the returned array was empty.

This caused a bug in the Flow component, as the Flow dataprovider
doesn't return the empty array again after it has already given this
data to the client. This left the combobox hanging in loading state.

Flow component issue:
https://github.com/vaadin/vaadin-combo-box-flow/issues/178