fix: bind value correctly when using attr-for-value
When using attr-for-value different than "value", the data binding of value wasn't fully working. This only happened to work with the iron-input example we have ("Replacing the Input" example) because iron-input also has a "value" property. When using a custom input with attr-for-value different than "value" the user was unable to change the value of the input by interacting with it via the keyboard. This is now also working.

Fixes #778