feat: add clear-button-visible API from <vaadin-text-field> (#794)
BREAKING CHANGE: The clear button is now hidden by default. To make it
visible, set the clear-button-visible attribute or use
the clearButtonVisible property.

* feat: add `clear-button-visible` API from `<vaadin-text-field>` (#794)

Fixes #786

* fix: focus, open/close and clear behaviour for `vaadin-combo-box-light`

There were some issues with `vaadin-combo-box-light` behaviour related to focus, opening/closing the overlay and using a custom clear button. Added `pointer-events: auto;` to `vaadin-combo-box-light` when the overlay is open, to make the basic click behaviour consistent with `vaadin-combo-box`.