Fix duplicate custom-value-set events fired
The _onClosed function was running twice when the overlay was closed:
- once by reacting to the dropdown closing event
- another time by the opened property observer

Because of this, the custom-value-set event was fired twice.

This was not caught by the tests, because the test combo box didn't
have any items, so there was no overlay, which would be closed to make
the extra _onClosed call.

Fix https://github.com/vaadin/vaadin-combo-box-flow/issues/167