test: fix flaky lazy-loading tests
The same Sinon spy instance was used as a dataProvider for all lazy-loading tests. This seemed to cause random failures (depending on how many tests are being run and in which browser) due to the spy being called again by combo box of a previous test fixture even though spy.reset() was called in beforeEach(). Another option to fix this would have been to unset the combobox dataProvider in afterEach(), but it seemed cleaner to just create a new spy instance for every test (should be more reliable).

Fixes #801