test: fix flaky test
This test was failing about 50% of the time on at least Firefox when running at least the whole keyboard.html test suite or all tests. Passed most of the time if running just this test or a smaller section from the keyboard.html suite. Seemed to fail also on Safari and Chrome in CI.