fix: scrolling quickly to the bottom shows blank items
When the dropdown was being scrolled quickly, some pages weren't being loaded and skipped in the end. It's fixed by resending the skipped pages after the scroller reset and filter those indexes which are close to current scroller position, so as to avoid unnecessary duplicated pages and loading indicator hanging.

Fixes: https://github.com/vaadin/vaadin-combo-box/issues/970